feature: refactor into separate view
